% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose function}
\usage{
bin_choose(trials, success)
}
\arguments{
\item{trials:}{how many times experiment repeated}

\item{success:}{how many successes out of trials}
}
\value{
the value of combinations
}
\description{
Finds choose value of trials and success
}
