% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial cumulative function}
\usage{
bin_cumulative(trials = 1, prob = 1)
}
\arguments{
\item{trials:}{how many times experiment repeated}

\item{prob:}{probability of success}
}
\value{
a list of the probabilities and cumulative probabilities of each success
}
\description{
Finds the cumulative probabilities of all successes
}
