% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials:}{how many times experiment repeated}

\item{prob:}{probability of success}
}
\value{
a list of the probabilities of each success
}
\description{
Finds the probabilities of all successes
}
