% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{binomial probability function}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{trials:}{how many times experiment repeated}

\item{success:}{how many successes out of trials}

\item{prob:}{probability of success}
}
\value{
the probability of number of successes
}
\description{
Finds the value based on success, trials and probability of binomial distribution
}
