% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial variable function}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials:}{how many times experiment repeated}

\item{prob:}{probability of success}
}
\value{
trial and probability variable in a list format
}
\description{
displays trial and probability
}
